Move `bids_dataset` fixture into `conftest.py`
